build: file and folder organizing
중첩된 파일구조를 정리하였습니다.
