+++ src/pages/PageMyRecipes.js
@@ -1,25 +1,25 @@
 import { useEffect } from "react";
 import styled from "styled-components";
 
-import Header from "../../components/Header";
-import Main from "../../components/Main";
-import Footer from "../../components/Footer";
+import Header from "../components/Header";
+import Main from "../components/Main";
+import Footer from "../components/Footer";
 
-function PageTips({ loginUserInfo, handleLogin }) {
+function PageMyRecipes({ loginUserInfo, handleLogin }) {
   useEffect(() => {
-    document.title = "Tips";
+    document.title = "MyRecipes";
   }, []);
 
   return (
     <Container>
       <Header loginUserInfo={loginUserInfo} handleLogin={handleLogin} />
-      <Main>Tips 화면입니다.</Main>
+      <Main>MyRecipes 화면입니다.</Main>
       <Footer />
     </Container>
   );
 }
 
-export default PageTips;
+export default PageMyRecipes;
 
 const Container = styled.div`
   height: 100%;
