+++ src/pages/PageRankings.js
@@ -1,25 +1,25 @@
 import { useEffect } from "react";
 import styled from "styled-components";
 
-import Header from "../../components/Header";
-import Main from "../../components/Main";
-import Footer from "../../components/Footer";
+import Header from "../components/Header";
+import Main from "../components/Main";
+import Footer from "../components/Footer";
 
-function PageMyRecipes({ loginUserInfo, handleLogin }) {
+function PageRankings({ loginUserInfo, handleLogin }) {
   useEffect(() => {
-    document.title = "MyRecipes";
+    document.title = "Rankings";
   }, []);
 
   return (
     <Container>
       <Header loginUserInfo={loginUserInfo} handleLogin={handleLogin} />
-      <Main>MyRecipes 화면입니다.</Main>
+      <Main>Rankings 화면입니다.</Main>
       <Footer />
     </Container>
   );
 }
 
-export default PageMyRecipes;
+export default PageRankings;
 
 const Container = styled.div`
   height: 100%;
