--- src/pages/PageRecipe/index.js
@@ -1,7 +1,0 @@
-import React from "react";
-
-const PageRecipe = () => {
-  return <div>test</div>;
-};
-
-export default test;
