+++ src/pages/PageSingleRecipe.js
@@ -1,25 +1,25 @@
 import { useEffect } from "react";
 import styled from "styled-components";
 
-import Header from "../../components/Header";
-import Main from "../../components/Main";
-import Footer from "../../components/Footer";
+import Header from "../components/Header";
+import Main from "../components/Main";
+import Footer from "../components/Footer";
 
-function PageRankings({ loginUserInfo, handleLogin }) {
+function PageOneRecipe({ loginUserInfo, handleLogin }) {
   useEffect(() => {
-    document.title = "Rankings";
+    document.title = "OneRecipe";
   }, []);
 
   return (
     <Container>
       <Header loginUserInfo={loginUserInfo} handleLogin={handleLogin} />
-      <Main>Rankings 화면입니다.</Main>
+      <Main>OneRecipes 화면입니다. 화면 구성 변경이 필요합니다!</Main>
       <Footer />
     </Container>
   );
 }
 
-export default PageRankings;
+export default PageOneRecipe;
 
 const Container = styled.div`
   height: 100%;
