+++ src/pages/PageTips.js
@@ -1,25 +1,25 @@
 import { useEffect } from "react";
 import styled from "styled-components";
 
-import Header from "../../components/Header";
-import Main from "../../components/Main";
-import Footer from "../../components/Footer";
+import Header from "../components/Header";
+import Main from "../components/Main";
+import Footer from "../components/Footer";
 
-function Home({ loginUserInfo, handleLogin }) {
+function PageTips({ loginUserInfo, handleLogin }) {
   useEffect(() => {
-    document.title = "Home";
+    document.title = "Tips";
   }, []);
 
   return (
     <Container>
       <Header loginUserInfo={loginUserInfo} handleLogin={handleLogin} />
-      <Main>Home 화면입니다.</Main>
+      <Main>Tips 화면입니다.</Main>
       <Footer />
     </Container>
   );
 }
 
-export default Home;
+export default PageTips;
 
 const Container = styled.div`
   height: 100%;
